Moved cafeCleanUp to dataCleaner
